# docs/openapi/levelup.v1.yaml
openapi: 3.1.0
info:
  title: LevelUp API
  version: 1.0.0-draft
  description: Contrato inicial futuro. A implementação HTTP completa não faz parte da primeira etapa.
servers:
  - url: https://localhost:5001/api/v1
paths:
  /auth/login:
    post:
      operationId: login
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
        '400': { $ref: '#/components/responses/ValidationProblem' }
        '401': { $ref: '#/components/responses/UnauthorizedProblem' }
        '429': { $ref: '#/components/responses/RateLimitedProblem' }
...
components:
  responses:
    ValidationProblem:
      description: Validation failed
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemContract' }
    UnauthorizedProblem:
      description: Authentication required
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemContract' }
    RateLimitedProblem:
      description: Rate limit exceeded
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemContract' }
...
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }
        rememberMe: { type: boolean, default: false }
    AuthenticationResponse:
      type: object
      required: [user]
      properties:
        user: { $ref: '#/components/schemas/CurrentUserResponse' }
    CurrentUserResponse:
      type: object
      required: [id, name, email, emailConfirmed, profileCompleted, onboardingCompleted]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        nickname: { type: [string, 'null'] }
        email: { type: string, format: email }
        emailConfirmed: { type: boolean }
        profileCompleted: { type: boolean }
        onboardingCompleted: { type: boolean }
...
    ProblemContract:
      type: object
      required: [title, status, code, correlationId]
      properties:
        type: { type: [string, 'null'], format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: [string, 'null'] }
        instance: { type: [string, 'null'] }
        code: { type: string }
        correlationId: { type: string }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }